JS Date => MySQL Date fix.
Do not rely on JS's .toLocaleString() to get the proper date.  It was 
causing an error being parsed into a date that MySQL could understand.  
It's winky wonkus but all we really needed to do was extract YYYY-MM-DD 
for the dates anyway.  So that's what happens in the API now.